Optimize js workflows and upgrade yarn (#4882)
- Upgrade to yarn 3
- Optimize js workflows and begin doing all js workspace tasks as part
of the main js action workflow
- Begin caching yarn dependencies in CI for much faster install times
- Begin caching turbo artifacts in CI to skip tasks when possible
- Move most js tasks into turbo commands
- Fossa has been failing silently for js workspaces in CI for some time.
`fossa analyze` doesn't exit 1 when detected projects fail and after
reading the docs it's not clear that this is supported in the cli. As a
result we need to change a few things:
- Only scan root yarn.lock where all dependencies are listed.
Configuration variations were tried to preserve fossa knowledge of which
dependencies belong to which packages but this is likely too complex for
fossa or it's not trivial to configure. Scanning all dependencies from
the root should achieve what we need in fossa.
- Skip fossa tasks when unrelated files change to speed up CI time using
dorny/paths-filter@v2 action

GitOrigin-RevId: 95ac1cc96b0ad839f2a17c0124258efe9fd711a8