Automatically update lock file in CI for PRs only (#126)
The yarn cache task provides a useful option to fail if installing the
deps would modify the lock file. This prevents CI from running against
different deps than what is explicit in the lock file and what we run
locally.

We have a couple scenarios where this causes frequent CI failures. It
happens in webdev PRs when the public js-sdk repo pushes package.json
changes due to new versions published from there. We can't copy over the
yarn.lock from the public repo - we have private deps in webdev that
don't exist in the public repo.

This PR updates the lock file automatically on PRs only - we still want
to fail on main if a change is made that would update the lock file
because it indicates that the tests haven't actually run correctly.

---------

Co-authored-by: Lightspark Eng <engineering@lightspark.com>