[js] Update lock file to resolve `fetch is not defined` (#5987)
Something in the current lock file on main causes `fetch is not defined`
when the wallet tests are actually rerun. Currently the cached runs are
being replayed on main so this is not visible. `fetch` is available
globally in Node as of v18 so this error doesn't really make sense.
Issue is resolved when the lock file is updated manually after deleting.

---------

Co-authored-by: Lightspark Eng <engineering@lightspark.com>
GitOrigin-RevId: 361d6b998d8ddee2742cf62c93a3c1cdb2945ab9