[site] Simplify / improve UI logic for InitAccount (#7859)
- Simplify a lot of the hook complexity that was previously in App into
the InitAccount router function (now called checkAll).
- checkAll can be called from any component via runInitAccount to
quickly fetch the account status and route to the proper place. The
store updates and navigation changes are only applied if the completed
run is equal to the latest triggered run.
- There can still only be one initNode process per mode at a time to
prevent huge numbers of polling requests each time checkAll is called.
This also prevents mutations to the node being made out of sync. If
InitAccount has a initNodePromise pending for the current mode it will
be reused.

GitOrigin-RevId: 71d83dcf22d59019014b3e6c41a9f1008331374d