[ui] Simplify toReactNodes and remove old routes type (#12431)
Previously we were using a generic type variable to determine which
typography props were allowed based on typography type.

The problem with that, as the tests didn’t catch, if you have an array
of objects like:
```
[
  { text: "sdjknf", typography: { type: "Body", bodyProp1: "sjdnf" } },
  { text: "sdjknf", typography: { type: "Headline", headlineProp1: "sjdnf" } }
]
```
it breaks down because both objects end up sharing the type variable, so
it expects the types passed in to conform to both.

The easier way of doing this is to have a union of typography props in a
certain way:
```
export type TypographyPropsWithoutContent = {
  [K in TypographyTypeKey]: { type: K } & Omit<
    ComponentProps<(typeof typographyMap)[K]>,
    "children" | "content"
  >;
}[TypographyTypeKey];
```
and typescript is able to infer which props are allowed based on the
type specified alone, without variables

So everywhere was updated to remove all those variables, and also the
problem described above about arrays is fixed
TextNode is now the only one that allows typography specified on the
node, because it’s the only one that needs it:
```
export type TextNode = {
  text: string;
  id?: string;
  onClick?: () => void;
  typography?: TypographyPropsWithoutContent;
};
```

The other nodes allow typography as a prop to the component itself, and
LinkNode, NextLinkNode, CurrencyAmountNode, etc all have these

So you can do
```
<Link text="sdlfn" typography={...} />
```
in addition to:
```
{ link: { text: "sdfjkn", typography: {...} } }
```
or with the node shorthand:
```
link({ text: "skdnf", typography: {...} })
```
Props are inferred automatically directly from the component, so any
prop can now be provided to the node

GitOrigin-RevId: 416c712b4557d377c1eb276706bb69e0abe5faf0