[private-ui] Move recaptcha to service (#13321)
Moves recaptcha to a service for better control of script loading
sequence and error handling. If recaptcha fails it surfaced as an error
to the user advising them to reload the page and tracked in frontend
logs. Instead of trying to run mutations with empty string as recaptcha
token this will fail early and display an error to the user.

We can be more clever about this in the future and try reloading the
script a few times if we continue seeing a lot of these errors in logs.
Let's monitor these changes first.

<img width="690" alt="Screenshot 2024-11-06 at 10 48 33 PM"
src="https://github.com/user-attachments/assets/42fdf777-3a2e-4688-8153-67b0cc07179d">
GitOrigin-RevId: 8516671aee3369231c498f468b01ac850c461a7e