[uma-bridge] Fix autofocus (#13502)
- iOS only allows autofocus (pulling up keyboard automatically on each
send screen) when the input is shown synchronously in response to a user
event. We can ease the UX in general by debouncing and running
`lookup_estimated_currency_conversions` while the user types so the
lnurlp is ready when user clicks "Continue". And this way the amount
input can be autofocused.

https://github.com/user-attachments/assets/9262db64-d008-4ecc-a43b-c41ce330dab7

GitOrigin-RevId: 4bb880ee550eddfb78b0cfe78b5932bac3a9e960