L1 wallet page (#13827)
Adds in the new page for the L1 wallet. Functionality for enabling and
disabling, and their visual states, is added in a subsequent PR. Uses Gk
to disabled access to the page until the feature is ready to launch.
<img width="599" alt="Screenshot 2024-12-16 at 9 38 21 AM"
src="https://github.com/user-attachments/assets/a68bdc9f-8127-4d0b-ab99-d0c048402dcb"
/>
GitOrigin-RevId: 96d6aa90ff2e0bc4a9f6f009e1c399dc4c745cdd