[lightspark-sdk] Handle double-encoded variables from WASM in RemoteSigningWebhookHandler (#19946)
## Reason

@pengying discovered that apps/remote-signing-server is failing due to
variables being sent to sparkcore as a string instead of a plain object.
We discovered this is because the variables returned from a WASM call is
double-encoded JSON. The code is not widely used which explains why we
haven't noticed until now.

## Overview

For large or complex changes, describe what is being changed.

## Test Plan

Tested manually and ensure no existing tests are broken.

GitOrigin-RevId: 347cb22a77fcc6f61fc3cffdd879078ac9d4982c