feat: adding idempotency to typescript client payInvoice (#20102)
## Reason

Added support for idempotency keys in the `payInvoice` method.  This was already available via the API but not exposed in the client lib.

## Overview

This change adds an optional `idempotencyKey` parameter to the `payInvoice` method in the Lightspark SDK. When provided, this key is passed to the GraphQL API, allowing the server to identify and prevent duplicate payment attempts.

GitOrigin-RevId: 78045ab188bddb612c6b19431419ac35f133c660