claude dot em dee (#20770)
## Reason

Claude.md! you can get these by running /init in the main repo

however, this pr splits it into **three** separate claude files.
the reason for this is that claude md files are recursive, so you can store javascript specific context in the js/claude.md, and it'll be added to context if you do any searches or tool calls in the /js directory

GitOrigin-RevId: 388a42b64bb640dcc522ba27b725683cae451b6d